Some library changes

+++ Desarrollo/Motor/Motor/EngineDevice.h
@@ -53,7 +53,7 @@ private:
 
 	int numFrames = 0;
 	int fps;
-	float screenWidth, screenHeight;
+	int screenWidth, screenHeight;
 
 	Input &input = Input::getInstance();
